// src/models/students.model.ts
export interface GetStudents
{
    id: string
    first_name: string
    last_name: string
    email: string
    age: number
    course: string
    year_level: number
    gpa: number
    enrollment_status: string
    created_at: string;
}

export interface CreateStudentPayload
{
    first_name: string
    last_name: string
    email: string
    age: number
    course: string
    year_level: number
    gpa: number
    enrollment_status: string
}